+++ src/scripts/hisenseTV/subscriptShutdownHandler.type.ts
@@ -1,0 +1,2 @@
+// returns exit code
+export type SubscriptExitCode = number;